src/requirements.txt
# Chassis del proyecto
microservice-chassis-grupo2_cc_prod

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# ML y procesamiento
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# HTTP client
httpx>=0.25.0

# Logging
python-json-logger>=2.0.0

# Base de datos (MySQL - el chassis usa aiomysql)
aiomysql>=0.2.0
sqlalchemy>=2.0.0
